Add support to optionally check for updates from GitHub
If the global constant `FEVER_GITHUB_REPOSITORY` is defined, Fever will attempt to request the latest release from the given GitHub repository via the GitHub REST API.

The constant `FEVER_GITHUB_REPOSITORY` is to the `<owner>/<repo>` to check.

The constant `FEVER_GITHUB_API_TOKEN ` is used to extend the rate limit.